+++ models/__init__.py
@@ -3,3 +3,4 @@ from . import issues
 from . import function
 from . import report_issues
 from . import report_project_list_xls
+from . import times
